Fix debug prints and register type in VM components

+++ src/vm/handlers/vm_control_flow_handlers.c
@@ -14,7 +14,7 @@
 #include "vm/vm_control_flow.h"
 
 // Ensure correct return type is known to the compiler
-extern Value vm_get_register_safe(uint8_t reg);
+extern Value vm_get_register_safe(uint16_t reg);
 
 // ====== Jump Operation Handlers ======
